backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==4.2
django-cors-headers==4.6.0
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==33.1.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
gprof2dot==2024.6.6
greenlet==3.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.4.2
locust==2.34.1
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pyzmq==26.4.0
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
toposort==1.10
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
